+++ packages/components-js/tests/a11y/specs/axe-core/common/components.a11y.ts
@@ -23,7 +23,7 @@ const revertAutoFocus = async (page: Page, component: string): Promise<void> => 
   }
 };
 
-const amountOfTestableComponents = 48;
+const amountOfTestableComponents = 49;
 
 test(`should have certain amount of components`, () => {
   expect(components.length).toBe(amountOfTestableComponents);