+++ packages/components/src/components/select-wrapper/select-wrapper-dropdown/select-wrapper-dropdown.tsx
@@ -235,10 +235,28 @@ export class SelectWrapperDropdown {
                   </li>
                 ) : (
                   this.optionMaps.map((option, index) => {
-                    const { value, disabled, hidden, initiallyHidden, selected, highlighted, title } = option;
+                    const {
+                      value,
+                      disabled,
+                      hidden,
+                      initiallyHidden,
+                      selected,
+                      highlighted,
+                      title,
+                      showOptgroup,
+                      optgroupOptions,
+                    } = option;
+
                     return [
-                      title && (
-                        <span class="optgroup" role="presentation">
+                      showOptgroup && (
+                        <span
+                          class={{
+                            optgroup: true,
+                            'optgroup--hidden': optgroupOptions.hidden,
+                            'optgroup--disabled': optgroupOptions.disabled,
+                          }}
+                          role="presentation"
+                        >
                           {title}
                         </span>
                       ),
@@ -251,6 +269,7 @@ export class SelectWrapperDropdown {
                           'option--highlighted': highlighted,
                           'option--disabled': disabled,
                           'option--hidden': hidden || initiallyHidden,
+                          'option--indent': !!title,
                         }}
                         onClick={!selected && !disabled ? () => this.setOptionSelected(index) : undefined}
                         {...getOptionAriaAttributes(selected, disabled, hidden, !!value)}