+++ packages/components/src/utils/validation/throwIfParentIsNotOfKind.ts
@@ -1,13 +1,20 @@
 import type { TagName } from '@porsche-design-system/shared';
-import { getPrefixedTagNames, getTagName, throwException } from '..';
-import { isParentOfKind } from '../dom/isParentOfKind';
+import {getPrefixedTagNames, getTagName, paramCaseToCamelCase, throwException} from '..';
+import { isParentOfKind } from '../dom';
 
-export const throwIfParentIsNotOfKind = (element: HTMLElement, tagName: TagName): void => {
-  if (element.parentElement && !isParentOfKind(element, tagName)) {
-    const allowedTagName = getPrefixedTagNames(element)[tagName];
-    const actualTagName = getTagName(element.parentElement);
-    throwException(
-      `parent HTMLElement of ${getTagName(element)} should be of kind ${allowedTagName} but got ${actualTagName}.`
-    );
+export const throwIfParentIsNotOfKind = (element: HTMLElement, tagNameOrNames: TagName | TagName[]): void => {
+  if (element.parentElement) {
+    const tagNamesArray = Array.isArray(tagNameOrNames) ? tagNameOrNames : [tagNameOrNames];
+    const matches = tagNamesArray.some((tagName) => isParentOfKind(element, tagName));
+
+    if (!matches) {
+      const allowedTagNames = tagNamesArray
+        .map((tagName) => getPrefixedTagNames(element)[paramCaseToCamelCase(tagName)])
+        .join(' | ');
+      const actualTagName = getTagName(element.parentElement);
+      throwException(
+        `parent HTMLElement of ${getTagName(element)} should be of kind ${allowedTagNames} but got ${actualTagName}.`
+      );
+    }
   }
 };